# openapi.yaml
openapi: 3.0.0
info:
  title: FEPY Backend API
  description: API del Sistema de Facturación Electrónica para Paraguay (Integración SIFEN)
  version: 1.0.0
servers:
  - url: /api
    description: Base URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /facturar/crear:
    post:
      summary: Crear y encolar factura electrónica
      tags: [Facturación]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON de la factura que se enviará al SIFEN.
              properties:
                param: 
                  type: object
                  properties:
                    ruc: { type: string, description: "RUC Emisor" }
                data: 
                  type: object
      responses:
        '202':
          description: Factura encolada para procesamiento asíncrono
  /facturar/empresa/{ruc}:
    get:
      summary: Info de la empresa emisora antes de facturar
      tags: [Facturación]
      parameters:
        - in: path
          name: ruc
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empresa encontrada

  # ========================================
  # GESTIÓN DE FACTURAS (INVOICES)
  # ========================================
...
  /invoices/{id}:
    get:
      summary: Obtener detalle de factura
      tags: [Invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
...
  /factura/estado/{id}:
    get:
      summary: Consultar estado (y queue status) de una factura
      tags: [Queue & Background]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Status

  # ========================================
  # EVENTOS SIFEN
  # ========================================
...